+++ esp_cam_sensor/sensors/ov3660/include/ov3660.h
@@ -1,0 +1,28 @@
+/*
+ * SPDX-FileCopyrightText: 2025 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+#pragma once
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include "esp_cam_sensor_types.h"
+#include "ov3660_types.h"
+
+#define OV3660_SCCB_ADDR   0x3c
+
+/**
+ * @brief Power on camera sensor device and detect the device connected to the designated sccb bus.
+ *
+ * @param[in] config Configuration related to device power-on and detection.
+ * @return
+ *      - Camera device handle on success, otherwise, failed.
+ */
+esp_cam_sensor_device_t *ov3660_detect(esp_cam_sensor_config_t *config);
+
+#ifdef __cplusplus
+}
+#endif